+++ packages/zent/src/form/formulr/listeners/set.tsx
@@ -1,4 +1,6 @@
 import { useMemo } from 'react';
+import { asapScheduler, merge } from 'rxjs';
+import { filter, observeOn } from 'rxjs/operators';
 import FormContext, { IFormContext, useFormContext } from '../context';
 import { FieldSetModel, isFieldSetModel } from '../models';
 import { useModelFromContext } from './use-model';
@@ -34,4 +36,23 @@ export function FieldSetValue({
     );
   }
   return null;
+}
+
+/**
+ * Because `FieldSetModel.prototype.registerChild` will be
+ * called inside `useMemo`, consume at next micro task queue
+ * to avoid react warning below.
+ *
+ * Cannot update a component from inside the function body
+ * of a different component.
+ */
+export function getFieldSetChildChangeObservable(
+  fieldSet: FieldSetModel,
+  name: string
+) {
+  const $ = merge(fieldSet.childRegister$, fieldSet.childRemove$).pipe(
+    observeOn(asapScheduler),
+    filter(change => change === name)
+  );
+  return $;
 }
